move auth code in a place where code gets executed
